+++ src/main/java/com/example/mainproject/controller/EquipmentModelController.java
@@ -24,7 +24,7 @@ public class EquipmentModelController {
 
     @PostMapping
     public ResponseEntity createEquipModel(@RequestBody EquipmentModelEntity equipmentModel,
-                                           @RequestParam Long equip_class_id){
+                                           @RequestParam Integer equip_class_id){
         try {
             return ResponseEntity.ok(equipmentModelService.createData(equipmentModel, equip_class_id));
         } catch (Exception e) {
