+++ src/main/java/com/example/mainproject/repository/ActionGroupRepo.java
@@ -3,5 +3,5 @@ package com.example.mainproject.repository;
 import com.example.mainproject.entity.ActionGroupEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ActionGroupRepo extends CrudRepository<ActionGroupEntity, Long> {
+public interface ActionGroupRepo extends CrudRepository<ActionGroupEntity, Integer> {
 }
