+++ src/main/java/com/example/mainproject/repository/ActionTypeRepo.java
@@ -1,7 +1,8 @@
 package com.example.mainproject.repository;
 
 import com.example.mainproject.entity.ActionTypeEntity;
+import jakarta.persistence.criteria.CriteriaBuilder;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ActionTypeRepo extends CrudRepository<ActionTypeEntity, Long> {
+public interface ActionTypeRepo extends CrudRepository<ActionTypeEntity, Integer> {
 }
