+++ src/main/java/com/example/mainproject/repository/DeptTypeRepo.java
@@ -3,5 +3,5 @@ package com.example.mainproject.repository;
 import com.example.mainproject.entity.DeptTypeEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DeptTypeRepo extends CrudRepository<DeptTypeEntity, Long> {
+public interface DeptTypeRepo extends CrudRepository<DeptTypeEntity, Integer> {
 }
