+++ src/main/java/com/example/mainproject/repository/EquipmentCategoryRepo.java
@@ -3,5 +3,5 @@ package com.example.mainproject.repository;
 import com.example.mainproject.entity.EquipmentCategoryEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EquipmentCategoryRepo extends CrudRepository<EquipmentCategoryEntity, Long> {
+public interface EquipmentCategoryRepo extends CrudRepository<EquipmentCategoryEntity, Integer> {
 }
