+++ src/main/java/com/example/mainproject/repository/EquipmentClassRepo.java
@@ -3,5 +3,5 @@ package com.example.mainproject.repository;
 import com.example.mainproject.entity.EquipmentClassEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EquipmentClassRepo extends CrudRepository<EquipmentClassEntity, Long> {
+public interface EquipmentClassRepo extends CrudRepository<EquipmentClassEntity, Integer> {
 }
