+++ src/main/java/com/example/mainproject/repository/HangerTypeRepo.java
@@ -3,5 +3,5 @@ package com.example.mainproject.repository;
 import com.example.mainproject.entity.HangerTypeEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface HangerTypeRepo extends CrudRepository<HangerTypeEntity, Long> {
+public interface HangerTypeRepo extends CrudRepository<HangerTypeEntity, Integer> {
 }
